move StatementSyntax extensions from components to syntax
