+++ packages/ember-glimmer/lib/syntax/curly-component.js
