+++ packages/ember-glimmer/lib/syntax/dynamic-component.js
